% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceDotplot.R
\name{abundanceDotplot}
\alias{abundanceDotplot}
\title{Relative Abundance Dotplot}
\usage{
abundanceDotplot(
  srt_obj,
  celltypes,
  group2,
  celltype_pal,
  use_facet = FALSE,
  plot_type = "dot"
)
}
\arguments{
\item{srt_obj}{Seurat object}

\item{celltypes}{Character string of cell type column in metadata}

\item{group2}{Character string of grouping column in metadata}

\item{celltype_pal}{Vector of colors for cell types}

\item{use_facet}{Logical for whether to use facet wrapping}

\item{plot_type}{Character string for plot type, either "dot" or "bar",
default is "dot"}
}
\value{
ggplot object
}
\description{
Relative Abundance Dotplot
}
\examples{
\dontrun{
  abundanceDotplot(seuratdata, celltypes = "celltypes", group2 = "orig.ident",
  celltype_pal = celltype_cols, use_facet = FALSE)
}
}
