% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCelltypeColorPal.R
\name{addCelltypeColorPal}
\alias{addCelltypeColorPal}
\title{Add celltype color palette to Seurat object}
\usage{
addCelltypeColorPal(srt_obj, celltypes)
}
\arguments{
\item{srt_obj}{Seurat object}

\item{celltypes}{character string of the column name in the metadata of the Seurat object that contains the celltype information}
}
\value{
celltype_cols a named vector of colors
}
\description{
This function adds a color palette to the metadata of a Seurat object based on the celltype column.
The function uses color palettes from khroma, ggthemes, and Rcolorbrewer.
https://github.com/tesselle/khroma
Frerebeau, N. (2024). khroma: Colour Schemes for Scientific Data Visualization. https://doi.org/10.5281/zenodo.1472077
https://github.com/jrnold/ggthemes
https://cran.r-project.org/web/packages/RColorBrewer/index.html
https://colorbrewer2.org/
}
\examples{
\dontrun{
srt_obj <- addCelltypeColorPal(srt_obj, "celltypes")
}
}
