% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPCs.R
\name{findPCs}
\alias{findPCs}
\title{Determine the number of PCs to use for downstream analysis}
\usage{
findPCs(
  seurat_object,
  threshold_pct = 5,
  threshold_cum_pct = 90,
  threshold_diff_pct = 0.1,
  elbow_plot = FALSE
)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{threshold_pct}{Threshold for percent of variation associated with each PC}

\item{threshold_cum_pct}{Threshold for cumulative percent of variation associated with each PC}

\item{threshold_diff_pct}{Threshold for difference between variation of PC and subsequent PC}

\item{elbow_plot}{Logical value to determine if an elbow plot should be generated}
}
\value{
Number of PCs to use for downstream analysis
}
\description{
This function was adapted from https://hbctraining.github.io/scRNA-seq/lessons/elbow_plot_metric.html
}
\examples{
findPCs(seurat_object,
        threshold_pct = 5,
        threshold_cum_pct = 90,
        threshold_diff_pct = 0.1,
        elbow_plot=TRUE)

}
