% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_percentage.R
\name{gene_percentage}
\alias{gene_percentage}
\title{Get percentage of gene expression greater than zero in dataset}
\usage{
gene_percentage(seurat_obj, gene)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{gene}{A gene of interest}
}
\value{
A printed message containing gene expression percentage
}
\description{
Get percentage of gene expression greater than zero in dataset
}
