% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getresolutionsizes.R
\name{getresolutionsizes}
\alias{getresolutionsizes}
\title{Number of unique clusters generated per resolution}
\usage{
getresolutionsizes(seurat_data, min_res = FALSE)
}
\arguments{
\item{seurat_data}{A Seurat object}

\item{min_res}{A logical indicating whether to identify resolution with minimum number of clusters}
}
\value{
Returns either the size of unique clusters generated or specific resolution with smallest unique cluster size
}
\description{
Number of unique clusters generated per resolution
}
