% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pals.R
\name{grabbag_pal}
\alias{grabbag_pal}
\title{A custom palette generator}
\usage{
grabbag_pal(pal_name, n)
}
\arguments{
\item{pal_name}{name of color palette. Options include:
\code{ucm_pal}, \code{ito_pal}, \code{tol_bright}, \code{tol_light},
\code{tol_muted}}

\item{n}{number of colors to output}
}
\value{
a vector containing desired colors
}
\description{
Code generates custom length color palettes from a collection of color blind
friendly palettes
}
\examples{
\dontrun{
grabbag_pal('ucm_pal', 2)
grabbag_pal('ito_pal', 8)
grabbag_pal('tol_muted', 10)
}

}
