% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pals.R
\docType{data}
\name{grabbag_pal_list}
\alias{grabbag_pal_list}
\title{List of my favorite color palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
grabbag_pal_list
}
\description{
Use \code{grabbag_pal} to pull palette of interest
}
\details{
Color palettes were collected from:
https://marketing.ucmerced.edu/resources/brand-guidelines/colors
https://thenode.biologists.com/data-visualization-with-flying-colors/research/
https://doi.org/10.1038/nmeth.1618
https://personal.sron.nl/~pault/
}
\keyword{datasets}
