% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepAnalysisDir.R
\name{prepAnalysisDir}
\alias{prepAnalysisDir}
\title{This function creates a directory for the sample and subdirectories for plots, data, and results.}
\usage{
prepAnalysisDir(sample_name)
}
\arguments{
\item{sample_name}{Name of the sample}
}
\value{
A directory for the sample and subdirectories for plots, data, and results
}
\description{
This function creates a directory for the sample and subdirectories for plots, data, and results.
}
\examples{
\dontrun{
prepAnalysisDir('sample1')
}
}
