% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runImmunePredictions.R
\name{runImmunePredictions}
\alias{runImmunePredictions}
\title{Run SingleR, scType and ProjectTILs on a Seurat object}
\usage{
runImmunePredictions(seurat_obj, cluster_id)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{cluster_id}{Cluster ID column name}
}
\value{
Data frame with SingleR, scType and ProjectTILs predictions
}
\description{
Run SingleR, scType and ProjectTILs on a Seurat object
}
\examples{

# do not add quotation marks around cluster_id
\dontrun{
runImmunePredictions(seurat_obj, seurat_cluster)
}

}
