% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSoupX.R
\name{RunSoupX}
\alias{RunSoupX}
\title{Run SoupX using Seurat clustering}
\usage{
RunSoupX(
  filtered_cellranger_counts_dir,
  raw_cellranger_counts_h5,
  output_sample_name,
  output_dir = NULL,
  output_type = "rds"
)
}
\arguments{
\item{filtered_cellranger_counts_dir}{Path to the filtered cellranger counts directory}

\item{raw_cellranger_counts_h5}{Path to the raw cellranger counts h5 file}

\item{output_sample_name}{Name of the output file}

\item{output_type}{Type of output file. Either "rds" or "mtx"}
}
\value{
A matrix of corrected counts
}
\description{
Run SoupX using Seurat clustering
}
\examples{
\dontrun{
RunSoupX(filtered_cellranger_counts_dir = "filtered_cellranger_counts_dir",
         raw_cellranger_counts_h5 = "raw_cellranger_counts_h5",
         output_sample_name = "output_sample_name",
         output_type = "rds")
}

}
