% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVolcanoPlot.R
\name{scVolcanoPlot}
\alias{scVolcanoPlot}
\title{Function to create volcanoplot from Seurat scRNAseq differential expression data}
\usage{
scVolcanoPlot(
  dex_df,
  fc_cutoff = 2,
  presto_de = FALSE,
  title = NULL,
  subtitle = NULL,
  caption = NULL
)
}
\arguments{
\item{dex_df}{Dataframe of differential expression data}

\item{fc_cutoff}{Fold change cutoff for plotting}

\item{presto_de}{Logical value to determine if the dataframe is from presto}

\item{title}{Title of the plot}

\item{subtitle}{Subtitle of the plot}

\item{caption}{Caption of the plot}
}
\value{
A volcanoplot of the differential expression data
}
\description{
Function to create volcanoplot from Seurat scRNAseq differential expression data
}
\examples{
\dontrun{
scVolcanoPlot(DEWilcox)
}

}
