% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedBarplot.R
\name{stackedBarplot}
\alias{stackedBarplot}
\title{This function creates a stacked barplot from a dataframe}
\usage{
stackedBarplot(df, group_label, xlabel = "")
}
\arguments{
\item{df}{dataframe}

\item{group_label}{column name of the dataframe to be used as grouping variable}

\item{xlabel}{x-axis label}
}
\value{
a stacked barplot
}
\description{
This function creates a stacked barplot from a dataframe
}
\examples{
stackedBarplot(iris, group_label = Species)
}
