% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topDEGBarplot.R
\name{topDEGBarplot}
\alias{topDEGBarplot}
\title{This function will create a barplot
of the top differentially expressed genes per cluster}
\usage{
topDEGBarplot(
  dex_df,
  n_degs = 25,
  presto = FALSE,
  fc_line = 0.25,
  title = paste0("Top ", n_degs, " DEGs")
)
}
\arguments{
\item{dex_df}{A dataframe of differentially expressed genes}

\item{n_degs}{The number of differentially expressed genes to plot per cluster}

\item{presto}{Logical value to determine if the dataframe is from presto}

\item{fc_line}{The fold change line to plot}
}
\value{
A barplot of the top differentially expressed genes per cluster
}
\description{
This function will create a barplot
of the top differentially expressed genes per cluster
}
\examples{
\dontrun{
topDEGBarplot(DEWilcox)
}

}
