% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_expression.R
\name{visualize_expression}
\alias{visualize_expression}
\title{Generate multiple plots to visualize gene expression}
\usage{
visualize_expression(dataset, genelist, density = FALSE, pal_large = FALSE)
}
\arguments{
\item{dataset}{A seurat data object}

\item{genelist}{A list of genes}

\item{density}{A logical indicating whether to plot kernel density estimates or not}

\item{pal_large}{A logical indicating whether to use a large palette 'Tableau 20' or not}
}
\value{
Multiple plots showing clustering, expression, and or kernel density estimates for a gene/genes
}
\description{
Generate multiple plots to visualize gene expression
}
\examples{

}
