% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genePercentage.R
\name{gene_percentage}
\alias{gene_percentage}
\title{Calculate the percentage of cells expressing a gene with counts above the median}
\usage{
gene_percentage(seurat_obj, gene)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{gene}{Gene name}
}
\value{
Percent of cells with gene expression > 0 and > median
}
\description{
Calculate the percentage of cells expressing a gene with counts above the median
}
\examples{
\dontrun{
gene_percentage(seurat_obj, "Cd8b")
}

}
